requirements.txt
django-cors-middleware==1.3.1
psycopg2==2.7.3.2
